Restructured everything
Restructured everything to use multiple C files. generators.h is now obsolete, as it contains no code. As it turns out, it's a very bad idea to initialize data in an .h file which is what I was doing, hence me moving all the data initialization to its own file, thereby removing all code from generators. It's still there for the time being so I can merge In-Progress to master automatically.